Add mixins for tabbed items components
Steps and Tabs items had a lot in common. Those are now lightweight.
