+++ src/components/steps/StepItem.vue
@@ -1,47 +1,21 @@
 <script>
+import TabbedChildMixin from '../../utils/TabbedChildMixin.js'
+
 export default {
     name: 'BStepItem',
+    mixins: [TabbedChildMixin],
     props: {
         step: String | Number,
-        label: String,
         type: String | Object,
-        icon: String,
-        iconPack: String,
         clickable: {
             type: Boolean,
             default: undefined
-        },
-        visible: {
-            type: Boolean,
-            default: true
         }
     },
     data() {
         return {
-            isActive: false,
-            transitionName: null,
+            elementClass: 'step-item',
             _isStepItem: true // Used internally by Step
-        }
-    },
-    methods: {
-        /**
-        * Activate step, alter animation name based on the index.
-        */
-        activate(oldIndex, index) {
-            this.transitionName = index < oldIndex
-                ? this.$parent.vertical ? 'slide-down' : 'slide-next'
-                : this.$parent.vertical ? 'slide-up' : 'slide-prev'
-            this.isActive = true
-        },
-
-        /**
-        * Deactivate step, alter animation name based on the index.
-        */
-        deactivate(oldIndex, index) {
-            this.transitionName = index < oldIndex
-                ? this.$parent.vertical ? 'slide-down' : 'slide-next'
-                : this.$parent.vertical ? 'slide-up' : 'slide-prev'
-            this.isActive = false
         }
     },
     created() {
@@ -50,37 +24,6 @@ export default {
             throw new Error('You should wrap bStepItem on a bSteps')
         }
         this.$parent.refreshSlots()
-    },
-    beforeDestroy() {
-        this.$parent.refreshSlots()
-    },
-    render(createElement) {
-        // if destroy apply v-if
-        if (this.$parent.destroyOnHide) {
-            if (!this.isActive || !this.visible) {
-                return
-            }
-        }
-        const vnode = createElement('div', {
-            directives: [{
-                name: 'show',
-                value: this.isActive && this.visible
-            }],
-            attrs: { 'class': 'step-item' }
-        }, this.$slots.default)
-        // check animated prop
-        if (this.$parent.animated) {
-            return createElement('transition', {
-                props: {
-                    'name': this.transitionName
-                },
-                on: {
-                    'before-enter': () => { this.$parent.isTransitioning = true },
-                    'after-enter': () => { this.$parent.isTransitioning = false }
-                }
-            }, [vnode])
-        }
-        return vnode
     }
 }
 </script>
